Refactor test_data.py - will be a good home for more classifier data tests
